requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# Evaluation
evaluate>=0.4.0
sacrebleu>=2.3.0

# Environment configuration
python-dotenv>=1.0.0

# Optional: Weights & Biases for experiment tracking
wandb>=0.15.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0

